NPCBots: Add 3 new command states: `.npcbot command follow only` - makes bots follow and nothing else, `.npcbot command nocast` - disabled all spellcasting, `.npcbot command nolongcast` - disabled non-instant spells.
(cherry picked from commit e2b31d13813513b562a4aef1cb71f76cb0416509)

+++ src/server/game/AI/NpcBots/botmgr.cpp
@@ -1857,13 +1857,13 @@ void BotMgr::ReviveAllBots()
         _reviveBot(itr->second);
 }
 
-void BotMgr::SendBotCommandState(uint8 state)
+void BotMgr::SendBotCommandState(uint32 state)
 {
     for (BotMap::const_iterator itr = _bots.begin(); itr != _bots.end(); ++itr)
         itr->second->GetBotAI()->SetBotCommandState(state, true);
 }
 
-void BotMgr::SendBotCommandStateRemove(uint8 state)
+void BotMgr::SendBotCommandStateRemove(uint32 state)
 {
     for (BotMap::const_iterator itr = _bots.begin(); itr != _bots.end(); ++itr)
         itr->second->GetBotAI()->RemoveBotCommandState(state);
