+++ src/server/game/AI/NpcBots/bpet_ai.h
@@ -44,9 +44,9 @@ class bot_pet_ai : public CreatureAI
         uint32 GetLastDiff() const { return lastdiff; }
         void CommonTimers(uint32 diff);
         void KillEvents(bool force);
-        void SetBotCommandState(uint8 st, bool force = false, Position* newpos = nullptr);
-        void RemoveBotCommandState(uint8 st);
-        bool HasBotCommandState(uint8 st) const { return (m_botCommandState & st); }
+        void SetBotCommandState(uint32 st, bool force = false, Position* newpos = nullptr);
+        void RemoveBotCommandState(uint32 st);
+        bool HasBotCommandState(uint32 st) const { return (m_botCommandState & st); }
         uint8 GetBotCommandState() const { return m_botCommandState; }
         bool IsInBotParty(Unit const* unit) const;
         virtual void ApplyBotPetSpellRadiusMods(SpellInfo const* /*spellInfo*/, float& /*radius*/) const {}
@@ -161,7 +161,7 @@ class bot_pet_ai : public CreatureAI
         static inline float _getAttackDistance(float distance) { return distance*0.72f; }
 
         Position movepos, attackpos;
-        uint8 m_botCommandState;
+        uint32 m_botCommandState;
 
         //timers
         uint32 lastdiff, checkAurasTimer, regenTimer, _updateTimerMedium, _updateTimerEx1;
